some more changes

+++ src/main/java/dev/dkvl/womutils/beans/MemberInfo.java
@@ -5,7 +5,7 @@ import lombok.Value;
 @Value
 public class MemberInfo
 {
-    int exp;
+    long exp;
     int id;
     String username;
     String displayName;
